+++ e2e-tests/tests/timers.rs
@@ -4,6 +4,12 @@ use std::time::Duration;
 mod test_utilities;
 use test_utilities::{cargo_build_canister, pic_base, update};
 
+/// Advance the time by a given number of seconds followed by a tick.
+fn advance_and_tick(pic: &PocketIc, seconds: u64) {
+    pic.advance_time(Duration::from_secs(seconds));
+    pic.tick();
+}
+
 #[test]
 fn test_timers() {
     let wasm = cargo_build_canister("timers");
@@ -13,24 +19,31 @@ fn test_timers() {
     pic.install_canister(canister_id, wasm, vec![], None);
 
     update::<(), ()>(&pic, canister_id, "schedule", ()).expect("Failed to call schedule");
-    advance_seconds(&pic, 5);
+    // without time skip
+    for _ in 0..5 {
+        advance_and_tick(&pic, 1);
+    }
+
+    update::<(), ()>(&pic, canister_id, "schedule", ()).expect("Failed to call schedule");
+    // with time skip: 3 and 4 are swapped because 3 is scheduled by another timer which must run first while 4 is available immediately
+    advance_and_tick(&pic, 5);
 
     update::<_, ()>(&pic, canister_id, "schedule_long", ()).expect("Failed to call schedule_long");
-    advance_seconds(&pic, 5);
+    advance_and_tick(&pic, 5);
     update::<_, ()>(&pic, canister_id, "cancel_long", ()).expect("Failed to call cancel_long");
-    advance_seconds(&pic, 5);
+    advance_and_tick(&pic, 5);
     update::<_, ()>(&pic, canister_id, "start_repeating", ())
         .expect("Failed to call start_repeating");
-    advance_seconds(&pic, 3);
+    advance_and_tick(&pic, 3);
     update::<_, ()>(&pic, canister_id, "stop_repeating", ())
         .expect("Failed to call stop_repeating");
-    advance_seconds(&pic, 2);
+    advance_and_tick(&pic, 2);
 
     let (events,): (Vec<String>,) =
         query_candid(&pic, canister_id, "get_events", ()).expect("Failed to call get_events");
     assert_eq!(
         events[..],
-        ["1", "2", "3", "4", "repeat", "repeat", "repeat"]
+        ["1", "2", "3", "4", "1", "2", "4", "3", "repeat", "repeat", "repeat"]
     );
 }
 
@@ -47,7 +60,7 @@ fn test_timers_can_cancel_themselves() {
     update::<_, ()>(&pic, canister_id, "set_self_cancelling_periodic_timer", ())
         .expect("Failed to call set_self_cancelling_periodic_timer");
 
-    advance_seconds(&pic, 1);
+    advance_and_tick(&pic, 1);
 
     let (events,): (Vec<String>,) =
         query_candid(&pic, canister_id, "get_events", ()).expect("Failed to call get_events");
@@ -61,7 +74,7 @@ fn test_timers_can_cancel_themselves() {
 fn test_scheduling_many_timers() {
     let wasm = cargo_build_canister("timers");
     // Must be more than the queue limit (500)
-    let timers_to_schedule = 1_000;
+    let timers_to_schedule = 1_000_u32;
     let pic = pic_base().build();
     let canister_id = pic.create_canister();
     pic.add_cycles(canister_id, 100_000_000_000_000u128);
@@ -77,20 +90,15 @@ fn test_scheduling_many_timers() {
 
     // Up to 20 timers will be executed per round
     // Be conservative that advance 2 times the minimum number of rounds
-    const TIMERS_PER_ROUND: u32 = 20;
-    advance_seconds(&pic, 2 * timers_to_schedule / TIMERS_PER_ROUND);
+    const TIMERS_PER_ROUND: u64 = 20;
+    for _ in 0..2 * timers_to_schedule as u64 / TIMERS_PER_ROUND {
+        advance_and_tick(&pic, 1);
+    }
 
     let (executed_timers,): (u32,) = query_candid(&pic, canister_id, "executed_timers", ())
         .expect("Error querying executed_timers");
 
     assert_eq!(timers_to_schedule, executed_timers);
-}
-
-fn advance_seconds(pic: &PocketIc, seconds: u32) {
-    for _ in 0..seconds {
-        pic.advance_time(Duration::from_secs(1));
-        pic.tick();
-    }
 }
 
 #[test]
@@ -108,7 +116,7 @@ fn test_set_global_timers() {
     update::<_, ()>(&pic, canister_id, "schedule_long", ()).expect("Failed to call schedule_long");
 
     // 5 seconds later, the 9s timer is still active
-    advance_seconds(&pic, 5);
+    advance_and_tick(&pic, 5);
 
     // Set the expiration time of the timer to t2 = t1 + 5s
     let t2 = t1 + 5_000_000_000;
@@ -132,7 +140,7 @@ fn test_async_timers() {
     pic.install_canister(canister_id, wasm, vec![], None);
 
     update::<(), ()>(&pic, canister_id, "async_await", ()).unwrap();
-    advance_seconds(&pic, 5);
+    advance_and_tick(&pic, 5);
 
     let (events,): (Vec<String>,) =
         query_candid(&pic, canister_id, "get_events", ()).expect("Failed to call get_events");
@@ -151,5 +159,27 @@ fn test_async_timers() {
             .filter(|e| *e == "method concurrent")
             .count(),
         3
+    );
+}
+
+#[test]
+fn test_periodic_timers_repeat_when_tasks_make_calls_despite_time_skipping() {
+    let wasm = cargo_build_canister("timers");
+    let pic = pic_base().build();
+
+    let canister_id = pic.create_canister();
+    pic.add_cycles(canister_id, 2_000_000_000_000);
+    pic.install_canister(canister_id, wasm, vec![], None);
+
+    update::<(), ()>(&pic, canister_id, "start_repeating_async", ()).unwrap();
+    // start_repeating_async sets a repeating timer with a 1s interval
+    // advance time by 3 seconds should trigger 3 repeats
+    advance_and_tick(&pic, 3);
+    update::<(), ()>(&pic, canister_id, "stop_repeating", ()).unwrap();
+
+    let (events,): (Vec<String>,) = query_candid(&pic, canister_id, "get_events", ()).unwrap();
+    assert_eq!(
+        events[..],
+        ["method repeat", "method repeat", "method repeat"]
     );
 }